Feat: Adicionado um novo robô, para capturar o input do usuário. O robô de texto já está pegando o conteúdo da Wikipedia e salvando no objeto content. Porém esse conteúdo não foi tratado.
