+++ .gitignore
@@ -59,3 +59,6 @@ typings/
 
 # next.js build output
 .next
+
+#credentials
+credentials/*.json
