+++ credentials/README.md
@@ -1,0 +1,10 @@
+# Credentials format
+
+## Algorithmia 
+
+File: `algorithmia.json`
+
+```
+    "ApiKey": "sim8As72h56j3mBYqEOSHfLE4k81"
+```
+Obs: esta credencial é a antiga do @filipedeschamps
