+++ package.json
@@ -23,6 +23,7 @@
   },
   "homepage": "https://github.com/williamtorres1/produtor-video#readme",
   "dependencies": {
+    "algorithmia": "^0.3.10",
     "readline-sync": "^1.4.10"
   }
 }
