+++ robots/user-input.js
@@ -1,0 +1,24 @@
+//biblioteca readline-sync importada no comando abaixo
+const readLine = require('readline-sync');
+
+function robot(content)
+{
+    content.searchTerm = askAndReturnSearchTerm();
+    content.prefix = askAndReturnPrefix();
+    
+    function askAndReturnSearchTerm()
+    {
+        return readLine.question('Type a Wikipedia search term: ');
+    }
+
+    function askAndReturnPrefix()
+    {
+        const prefixes = ['Who is', 'What is', 'The history of'];
+        const selectedPrefixIndex = readLine.keyInSelect(prefixes, 'Choose one option: ');
+        const selectedPrefixText = prefixes[selectedPrefixIndex];
+
+        return selectedPrefixText;
+    }
+}
+
+module.exports = robot;
